+++ common/src/main/java/club/iananderson/seasonhud/client/gui/components/sliders/BasicSlider.java
@@ -77,7 +77,7 @@ public class BasicSlider extends AbstractSliderButton {
   protected BasicSlider(int x, int y, int width, int height, boolean drawString, double initial, double minValue,
       double maxValue, double defaultValue, double stepSize, int precision) {
     this(x, y, width, height, drawString, initial, minValue, maxValue, defaultValue, stepSize, precision,
-        ChatFormatting.WHITE);
+         ChatFormatting.WHITE);
   }
 
   protected BasicSlider(int x, int y, int width, int height, boolean drawString, double initial, double minValue,
@@ -260,9 +260,9 @@ public class BasicSlider extends AbstractSliderButton {
   @Override
   protected void renderBg(@NotNull PoseStack graphics, @NotNull Minecraft mc, int mouseX, int mouseY) {
     DrawUtil.blitWithBorder(graphics, this, SLIDER_LOCATION, this.x, this.y, 0, this.getTextureY(), this.width,
-        this.height, 200, 20, 2, 3, 2, 2);
+                            this.height, 200, 20, 2, 3, 2, 2);
     DrawUtil.blitWithBorder(graphics, this, SLIDER_LOCATION, this.x + (int) (this.value * (double) (this.width - 8)),
-        this.y, 0, this.getHandleTextureY(), 8, this.height, 200, 20, 2, 3, 2, 2);
+                            this.y, 0, this.getHandleTextureY(), 8, this.height, 200, 20, 2, 3, 2, 2);
     this.renderScrollingString(graphics, mc.font, 2, this.getFgColor() | Mth.ceil(this.alpha * 255.0F) << 24);
   }
 }