+++ common/src/main/java/club/iananderson/seasonhud/client/gui/components/sliders/rgb/RgbSlider.java
@@ -4,7 +4,7 @@ import club.iananderson.seasonhud.Common;
 import club.iananderson.seasonhud.client.gui.components.boxes.ColorEditBox;
 import club.iananderson.seasonhud.client.gui.components.sliders.BasicSlider;
 import club.iananderson.seasonhud.config.SeasonHudClient;
-import club.iananderson.seasonhud.impl.seasons.Seasons;
+import club.iananderson.seasonhud.impl.season.components.Seasons;
 import club.iananderson.seasonhud.util.Rgb;
 import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.ChatFormatting;
@@ -25,7 +25,7 @@ public class RgbSlider extends BasicSlider {
 
   public RgbSlider(int x, int y, int initial, ColorEditBox seasonBox, ChatFormatting textColor) {
     super(x, y, seasonBox.getWidth() + 2, seasonBox.getHeight() - 6, true, initial, 0, 255,
-        seasonBox.getSeason().getDefaultColor(), textColor);
+          seasonBox.getSeason().getDefaultColor(), textColor);
     this.seasonBox = seasonBox;
     this.season = seasonBox.getSeason();
     this.rgb = Integer.parseInt(seasonBox.getValue());