+++ common/src/main/java/club/iananderson/seasonhud/client/gui/screens/MainConfigScreen.java
@@ -4,14 +4,12 @@ import club.iananderson.seasonhud.Common;
 import club.iananderson.seasonhud.client.gui.components.buttons.MenuButton;
 import club.iananderson.seasonhud.client.gui.components.buttons.MenuButton.MenuButtons;
 import club.iananderson.seasonhud.config.SeasonHudClient;
-import club.iananderson.seasonhud.impl.minimaps.CurrentMinimap;
+import club.iananderson.seasonhud.impl.minimap.CurrentMinimap;
 import club.iananderson.seasonhud.platform.Services;
 import com.mojang.blaze3d.vertex.PoseStack;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.List;
 import net.minecraft.client.gui.GuiComponent;
-import net.minecraft.client.gui.components.AbstractWidget;
 import net.minecraft.client.gui.components.CycleButton;
 import net.minecraft.network.chat.Component;
 import org.jetbrains.annotations.NotNull;
@@ -20,7 +18,6 @@ public class MainConfigScreen extends SeasonHudScreen {
   private static final Component SCREEN_TITLE = Common.translatedText("menu.seasonhud.main.title");
   private static final Component MINIMAP_SETTINGS = Common.translatedText("menu.seasonhud.main.minimap.options");
   private static final Component JOURNEYMAP = Common.translatedText("menu.seasonhud.main.journeymap.title");
-  private final List<AbstractWidget> optionButtons = new ArrayList<>();
   MenuButton seasonButton;
   MenuButton colorButton;
   CycleButton<Boolean> enableMinimapIntegrationButton;
@@ -96,6 +93,7 @@ public class MainConfigScreen extends SeasonHudScreen {
     showMinimapHiddenButton.active = enableMod;
   }
 
+  @SuppressWarnings("ConstantValue")
   @Override
   public void init() {
     super.init();
@@ -105,20 +103,18 @@ public class MainConfigScreen extends SeasonHudScreen {
     CycleButton<Boolean> enableModButton = CycleButton.onOffBuilder(enableMod)
         .withTooltip(t -> Common.newTooltip("menu.seasonhud.main.enableMod.tooltip"))
         .create(this.width - enableModWidth - TITLE_PADDING / 2, TITLE_PADDING / 2, enableModWidth, buttonHeight,
-            Common.translatedText("menu.seasonhud.main.enableMod.button"), (b, val) -> enableMod = val);
+                Common.translatedText("menu.seasonhud.main.enableMod.button"), (b, val) -> enableMod = val);
     widgets.add(enableModButton);
 
     int row = 0;
     seasonButton = MenuButton.builder(MenuButtons.SEASON, this, SeasonOptionsScreen.getInstance(this))
         .withTooltip(Common.newTooltip("menu.seasonhud.main.season.tooltip"))
-        .withPos(leftButtonX, (buttonStartY + (row * offsetY)))
-        .withWidth(buttonWidth)
+        .withPos(leftButtonX, (buttonStartY + (row * offsetY))).withWidth(buttonWidth)
         .build();
 
     colorButton = MenuButton.builder(MenuButtons.COLORS, this, ColorScreen.getInstance(this))
         .withTooltip(Common.newTooltip("menu.seasonhud.main.color.tooltip"))
-        .withPos(rightButtonX, (buttonStartY + (row * offsetY)))
-        .withWidth(buttonWidth)
+        .withPos(rightButtonX, (buttonStartY + (row * offsetY))).withWidth(buttonWidth)
         .build();
     widgets.addAll(Arrays.asList(seasonButton, colorButton));
 
@@ -126,13 +122,14 @@ public class MainConfigScreen extends SeasonHudScreen {
     enableMinimapIntegrationButton = CycleButton.onOffBuilder(enableMinimapIntegration)
         .withTooltip(t -> Common.newTooltip("menu.seasonhud.main.minimapIntegration.tooltip"))
         .create(leftButtonX, (buttonStartY + (row * offsetY)), buttonWidth, buttonHeight,
-            Common.translatedText("menu.seasonhud.main.enableMinimapIntegration.button"),
-            (b, val) -> enableMinimapIntegration = val);
+                Common.translatedText("menu.seasonhud.main.enableMinimapIntegration.button"),
+                (b, val) -> enableMinimapIntegration = val);
 
     showMinimapHiddenButton = CycleButton.onOffBuilder(showMinimapHidden)
         .withTooltip(t -> Common.newTooltip("menu.seasonhud.main.showMinimapHidden.tooltip"))
         .create(rightButtonX, (buttonStartY + (row * offsetY)), buttonWidth, buttonHeight,
-            Common.translatedText("menu.seasonhud.main.showMinimapHidden.button"), (b, val) -> showMinimapHidden = val);
+                Common.translatedText("menu.seasonhud.main.showMinimapHidden.button"),
+                (b, val) -> showMinimapHidden = val);
 
     widgets.addAll(Arrays.asList(enableMinimapIntegrationButton, showMinimapHiddenButton));
 
@@ -141,13 +138,14 @@ public class MainConfigScreen extends SeasonHudScreen {
       journeyMapAboveMapButton = CycleButton.onOffBuilder(journeyMapAboveMap)
           .withTooltip(t -> Common.newTooltip("menu.seasonhud.main.journeymap.aboveMap.tooltip"))
           .create(leftButtonX, (buttonStartY + (row * offsetY)), buttonWidth, buttonHeight,
-              Common.translatedText("menu.seasonhud.main.journeymap.aboveMap.button"),
-              (b, val) -> journeyMapAboveMap = val);
+                  Common.translatedText("menu.seasonhud.main.journeymap.aboveMap.button"),
+                  (b, val) -> journeyMapAboveMap = val);
 
       journeyMapMacOsButton = CycleButton.onOffBuilder(journeyMapMacOs)
           .withTooltip(t -> Common.newTooltip("menu.seasonhud.main.journeymap.macOS.tooltip"))
           .create(rightButtonX, (buttonStartY + (row * offsetY)), buttonWidth, buttonHeight,
-              Common.translatedText("menu.seasonhud.main.journeymap.macOS.button"), (b, val) -> journeyMapMacOs = val);
+                  Common.translatedText("menu.seasonhud.main.journeymap.macOS.button"),
+                  (b, val) -> journeyMapMacOs = val);
 
       widgets.addAll(Arrays.asList(journeyMapAboveMapButton, journeyMapMacOsButton));
     }