+++ fabric/src/main/java/club/iananderson/seasonhud/fabric/event/ClientEvents.java
@@ -1,10 +1,10 @@
 package club.iananderson.seasonhud.fabric.event;
 
+import club.iananderson.seasonhud.SeasonHudClientCommon;
 import club.iananderson.seasonhud.client.KeyBindings;
-import club.iananderson.seasonhud.client.gui.screens.MainConfigScreen;
 import club.iananderson.seasonhud.fabric.client.overlays.JourneyMap;
 import club.iananderson.seasonhud.fabric.client.overlays.SeasonHudOverlay;
-import club.iananderson.seasonhud.impl.minimaps.CurrentMinimap;
+import club.iananderson.seasonhud.impl.minimap.CurrentMinimap;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper;
 
@@ -15,15 +15,13 @@ public class ClientEvents {
   // Key Bindings
   private static void registerKeyInputs() {
     ClientTickEvents.END_CLIENT_TICK.register(client -> {
-      if (KeyBindings.seasonhudOptionsKeyMapping.consumeClick()) {
-        MainConfigScreen.getInstance().open();
-      }
+      SeasonHudClientCommon.optionsKeyInput();
     });
   }
 
   private static void registerKeyMappings() {
-    KeyBindings.seasonhudOptionsKeyMapping =
-        KeyBindingHelper.registerKeyBinding(KeyBindings.seasonhudOptionsKeyMapping);
+    KeyBindings.seasonhudOptionsKeyMapping = KeyBindingHelper.registerKeyBinding(
+        KeyBindings.seasonhudOptionsKeyMapping);
   }
 
   private static void registerHud() {