+++ forge/src/main/java/club/iananderson/seasonhud/forge/SeasonHudForge.java
@@ -4,6 +4,7 @@ import club.iananderson.seasonhud.Common;
 import club.iananderson.seasonhud.config.SeasonHudClient;
 import club.iananderson.seasonhud.config.SeasonHudServer;
 import club.iananderson.seasonhud.forge.impl.curios.CuriosCompat;
+import club.iananderson.seasonhud.impl.minimap.CurrentMinimap;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.ModLoadingContext;